feat(emsch/sso): add AbstractOAuth2Provider.php
