+++ EMS/client-helper-bundle/src/Security/Sso/OAuth2/OAuth2Service.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace EMS\ClientHelperBundle\Security\Sso\OAuth2;
 
 use EMS\ClientHelperBundle\Controller\Security\Sso\OAuth2Controller;
-use EMS\ClientHelperBundle\Security\Sso\OAuth2\Provider\AzureProvider;
-use EMS\ClientHelperBundle\Security\Sso\OAuth2\Provider\KeycloakProvider;
+use EMS\ClientHelperBundle\Security\Sso\OAuth2\Provider\AzureOAuth2Provider;
+use EMS\ClientHelperBundle\Security\Sso\OAuth2\Provider\KeycloakOAuth2Provider;
 use EMS\ClientHelperBundle\Security\Sso\OAuth2\Provider\ProviderInterface;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\HttpFoundation\RedirectResponse;
@@ -87,14 +87,14 @@ class OAuth2Service
     private function createProvider(): ProviderInterface
     {
         return match ($this->property(OAuth2Property::PROVIDER)) {
-            'azure' => new AzureProvider(
+            'azure' => new AzureOAuth2Provider(
                 tenant: $this->property(OAuth2Property::REALM),
                 clientId: $this->property(OAuth2Property::CLIENT_ID),
                 clientSecret: $this->property(OAuth2Property::CLIENT_SECRET),
                 redirectUri: $this->property(OAuth2Property::REDIRECT_URI),
                 version: $this->optionalProperty(OAuth2Property::VERSION),
             ),
-            'keycloak' => new KeycloakProvider(
+            'keycloak' => new KeycloakOAuth2Provider(
                 authServerUrl: $this->property(OAuth2Property::AUTH_SERVER),
                 realm: $this->property(OAuth2Property::REALM),
                 clientId: $this->property(OAuth2Property::CLIENT_ID),
